refactor(oficina): remove shim MercosulPlate — consumidores direto no canon shared/ (ADR 0251 cumprida, ADR 0272 roadmap)
3 imports atualizados (ServiceOrderRichSheet, Board, ServiceOrderKanbanCard).
Gates: reuse ✅ 21 · eslint ✅ · layout ✅ · components ✅ · zero TS novo.
CHANGELOG DS 0.6.14. Zero visual.
