--- resources/js/Pages/OficinaAuto/ProducaoOficina/_components/MercosulPlate.tsx
@@ -1,6 +1,0 @@
-// MercosulPlate — PROMOVIDO pra @/Components/shared/MercosulPlate (ADR 0251).
-//
-// Este arquivo agora é só um re-export pra não quebrar os imports existentes do
-// OficinaAuto (ServiceOrderRichSheet, ServiceOrderKanbanCard).
-// Fonte canônica única lá — Sells (venda com veículo) usa a MESMA plaquinha.
-export { default } from '@/Components/shared/MercosulPlate';
